pyproject.toml
[tool.poetry]
name = "django-custom-admin"
version = "0.1.0"
description = ""
authors = ["Bitboy7 <luisganerxgmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-compressor = "^4.5.1"
mysqlclient = "^2.2.6"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
markuppy = "^1.14"
reportlab = "^4.2.5"
django-import-export = "^4.3.3"
pandas = "^2.2.3"
pillow = "^11.1.0"
django-unfold = "^0.62.0"
numpy = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
